refactor:  - move src/data-access -> src/database-server/data-access  - move src/data-model -> src/database-server/data-model
ref #11
