+++ src/database-server/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/data-model)
+add_subdirectory(${CMAKE_CURRENT_SOURCE_DIR}/data-access)
